Fix NativeLinkable deps for SwiftLibrary
Summary:
In the diff `First pass at Swift module and binary support`, we changed
`StringArg` to take an optional list of deps to include with the args
of the `NativeLinkableInput` returned by `SwiftLibrary.getNativeLinkableInput()`.

Turns out this isn't the correct way to add a dependency, due to the
way input-based rule keys work. This caused Buck to fail to re-link a
binary which depended on a `swift_library` whose `srcs` had changed.

For build rules like Swift which have multiple outputs, the correct
way is to construct a `BuildTargetSourcePath` using the two-arg
constructor to refer to a sub-`Path` under the build rule's output.

This reverts the change to `StringArg` and uses a `BuildTargetSourcePath`
instead to include the dependency on the file underneath the swift library's
build directory.

Test Plan:
New integration test added. Confirmed test failed before
change with `ant java-test -Dtest.class=SwiftOSXBinaryIntegrationTest`,
applied fix, re-ran test, confirmed test passed.

Reviewed By: bolinfest

fb-gh-sync-id: 2380016
shipit-source-id: 2380016