Make some uses of types explicit so they appear as imports.
Summary:
The upcoming `buck autodeps` feature performs much better when the types
that are used in a library appear as `import` statements in the code. It has
some heuristics and configuration options to work around dependencies that
do not appear as `import` statements, but those are generally messier than doing
the trick you see here where we turn implicit dependencies into explicit ones.

I am in the process of writing up the documentation for `buck autodeps`, so I
would like to get this committed and exported to GitHub so that I can link to it
from the docs.

A more detailed explanation of why this is necessary will be present in the
upcoming docs on `buck autodeps`.

Test Plan: Landcastle

Reviewed By: marcinkosiba

fb-gh-sync-id: 109e4b3
shipit-source-id: 109e4b3