Prefer ListeningExecutorService to ExecutorService.
Summary:
Previous to this change, `ExecutionContext#getExecutorService()` returned
an `ExecutorService`. This meant that every caller who wanted a
`ListeningExecutorService` would have to do:

```
ListeningExecutorService executorService = MoreExecutors.listeningDecorator(
    executionContext.getExecutorService(ExecutorPool.CPU));
```

Rather than place the burden on every caller to create a new decorator,
let's just do it in one place when the `ExecutorService` is added to the
`ExecutionContext.Builder`.

Test Plan: Sandcastle

Reviewed By: Coneko

fb-gh-sync-id: 0a5dc33
shipit-source-id: 0a5dc33