Fix Windows breakage in buck.py
Summary:
Thanks to janicduplessis on Github for reporting this!

When I converted buck.py to speak BSER instead of JSON,
I broke Windows compatibility.

The problem was that the Python `open` and `fdopen` APIs
open the file in "text" mode by default, which translates
`\r\n` to `\n` and vice-versa.

For binary files which don't use `\r\n` and `\n` to mean "end-of-line",
you have to manually pass the `b` flag to `open` and `fdopen`.

This adds the flag to the appropriate calls in `buck.py`.

Partial fix for https://github.com/facebook/buck/issues/649

Test Plan:
Tested by janicduplessis:
https://github.com/facebook/buck/issues/649#issuecomment-182153688

Reviewed By: Coneko, sdwilsh

fb-gh-sync-id: 4b4afd0
shipit-source-id: 4b4afd0