Introduce simple two-level cache support.
Summary:
When a new .buckversion is introduced it is very common for the dircache to already contain
the artifacts needed to rebuild. Because the cache associates artifact contents directly with
the RuleKeys this results in having to fetch the artifacts again.
Introducing an intermediate content hash mapping (RuleKey -> content hash -> artifact) we
address the situation at the expense of doubling the latency of a single fetch.

Test Plan: tested on the dir cache

Reviewed By: Coneko

fb-gh-sync-id: 94a3517
shipit-source-id: 94a3517