Easy: Make FakeWatchmanClient more resistant to misconfiguration
Summary:
Previously, `FakeWatchmanClient` would return `Optional.absent()`
if asked to perform a query for which it wasn't configured.

This return value actually means "timed out", so it was too easy to
accidentally write tests which passed an incorrect query.

This diff simply makes `FakeWatchmanClient` throw an exception whenever
passed a query for which it wasn't configured.

Test Plan:
  ant java-test -Dtest.class=WatchmanTest
  ant java-test -Dtest.class=WatchmanWatcherTest

Reviewed By: marcinkosiba

fb-gh-sync-id: 065cb01
shipit-source-id: 065cb01