Don't crawl repository twice looking for BUCK files
Summary:
Whenever running a command on an entire tree (say, `buck project`
or `buck build path/...`), we were invoking `BuildFileSpec.findBuildFiles()`
twice.

This method crawls the entire filesystem (slowly), and on a large repository on an HFS+
filesystem, it can take upwards of 30 seconds each time.

We don't need to crawl the filesystem twice, so this adds a cache to
remember the results of the crawl.

I also added a new perf event to confirm this is very slow. A subsequent
diff will improve the perf of this codepath.

Test Plan:
`buck build ...` in large repo, confirmed in trace that the new
`find_build_files` perf event took over 30 seconds of wall clck time

Reviewed By: sdwilsh

fb-gh-sync-id: 1c9a203
shipit-source-id: 1c9a203