Fix OnDiskMavenDownloader path not resolving properly on windows
OnDiskMavenDownloader#fetch used root.resolve to convert the uri returned
from the MavenUrlDecoder to a local file scheme. This did not work on
Windows because of the drive letter.
```
file:///C:/com/android/support/support-v4/23.0.1/support-v4-23.0.1.aar
```
on windows vs
```
file:///com/android/support/support-v4/23.0.1/support-v4-23.0.1.aar
```
on mac/linux

This removes the extra uri conversion that was breaking the path because it is only used to log the DownloadEvent which can use the actual target path instead.

Tested by downloading support-v4 while building react-native on both mac
and windows.