Introduce generated_symbols argument for java_library() and add immutable_srcs argument for java_immutables_library() macro.
Summary:Previous to this change, we had to provide a ton of configuration in `.buckconfig` for
the benefit of `buck autodeps` because it does not know about `.java` files that
will be generated by the annotation processor of a `java_library()` rule.

In practice, these come from types with `BuckStyleImmutable` that,
by convention, start with the name `Abstract`. In this diff, we exploit this convention
to make a simple change to the `java_immutables_library()` macro where the
generating sources are listed via a separate argument:

```
java_immutables_library(
  name = 'packageable',
  srcs = [
    'AndroidPackageable.java',
    'AndroidPackageableCollector.java',
  ],
  immutable_srcs = [
    'AbstractAndroidPackageableCollection.java',
  ],
```

This makes it possible to eliminate the following from `.buckconfig`:

```
java-package-mappings = \
    com.facebook.buck.android.AndroidPackageableCollection => //src/com/facebook/buck/android:packageable, \
```

All of the `BUCK` files have been updated to leverage the new argument in the macro
such that all of the related configuration has been removed from `.buckconfig`.

Moving this information closer to where it is defined should result in fewer
surprises for developers.

Test Plan:* With this change, ran `buck autodeps` and verified none of the `BUCK.autodeps` files were changed.
* CI
* Previewed docs at http://localhost:9811/rule/java_library.html

Reviewed By: marcinkosiba

fb-gh-sync-id: 8acfe69
shipit-source-id: 8acfe69