Update to pywatchmen v1.3.0 (v4.4.0 of watchman)
Summary:We were in a weird state here.  The files are really from watchman, although the
`README.facebook` file indicated they were from a project that just copied them.
When I updated to the right project, I noticed the license was different.  The
code is technically licensed for both, but I choose to list the watchman
license now.

Test Plan: CI

Reviewed By: bhamiltoncx

fb-gh-sync-id: 5fc5fa6
shipit-source-id: 5fc5fa6