Always ignore the trash dir
Summary:We do not always ignore buck-out like the comment in the code says.
Unfortunately, tracking that down doesn't look trivial, so I'm
papering over the issue I'm about to describe by ignoring the `.trash`
directory for now and adding a TODO.

When we noticed a changed in the `buckversion`, we move the contents
of `buck-out` into `buck-out/.trash` and then start to asynchronously
delete it.  However, we also start walking the filesystem early on
in order to find build files.  Since `buck-out` isn't actually
ignored, we end up walking it.  On Windows at least, we end up
getting a security exception because we are trying to walk a file
that was asynchronously removed.

Test Plan:It's hard to prove that something that happens intermittently is
actually fixed, but I'm not seeing it anymore.

Reviewed By: Coneko

fb-gh-sync-id: c2a9b1d
shipit-source-id: c2a9b1d