Allow dir cache to use file without copying it so speed things up
Summary:In some cases, dir cache is the only final user of the zip file that has been prepared by the build system.
In that case, we can avoid copying it to temp file and moving to the dir cache folder. We can move it to the folder
straight away without the fear that cache consistency will be broken. This hugely improves the performance on local
builds where we don't use http cache to write stuff, so there is only one read-write cache - dir cache.

Test Plan:CI
Measure build times locally

Reviewed By: ruibm

fb-gh-sync-id: acc08d6
shipit-source-id: acc08d6