split native and bytecode linkable inputs (take 2)
Summary:native inputs were being passed around as inputs to both the native and bytecode rules, and converted by regex.

this diff instead threads bytecode inputs through to the bytecode rules and native to native (as before). this eliminates the restriction that native and bytecode have the same filename (and in a followup diff, that native even exists at all).

(take 2: this was reverted because native inputs were being used in bytecode rules and this split exposed it. that underlying issue has been fixed)

Test Plan: tests

Reviewed By: int3

fb-gh-sync-id: 7a5c23b
shipit-source-id: 7a5c23b