Import test utility class FakeExecutor into Buck from jcommon
Summary:I'm fixing a bug in `CounterRegistryImpl` related to its
use of a `ScheduledExecutorService`.

The tests for this class currently use `EasyMock` to model the behavior
of `ScheduledExecutorService`. I don't really like `EasyMock`, so this
diff introduces a new fake class `FakeExecutor` which we can use as
a stand-in for `ScheduledExecutorService`.

This class came from `com.facebook.testing.MockExecutor` (plus its
dependencies) from https://github.com/facebook/jcommon/ .

The only changes I made were:

1) Name them "Fake" instead of "Mock"
2) Remove the dependency on slf4j
3) Fix compilation warnings due to underspecified generics
4) Update the copyright date

I looked at pulling in the entire jcommon testing jar as-is,
but it has a bunch of dependencies we don't care about, so it
seemed overkill.

Test Plan: Used in subsequent diff's tests

Reviewed By: ruibm

fb-gh-sync-id: e1a642b
shipit-source-id: e1a642b