Added interfaces and modified events
Summary:This diff separates the webserver events from the rest of the code, in order to be used by other external applications.
To do this we created interfaces for each of the used events.
This interfaces can be found in "src/com/facebook/buck/webserver/events".
Because we also need some more other elements as "TestCaseSummaryInterface", in order to not depend on a lot of things in buck and be able to use them in other applications we have this structures in "src/com/facebook/buck/webserver/elements".

This doesn't add any functionality change.
It just renames a couple of methods.

Test Plan: Ran the tests locally.

Reviewed By: marcinkosiba

fb-gh-sync-id: e898432
shipit-source-id: e898432