Introduce CxxToolFlags, which allows (platform, rule) tuples to be composed
Summary:Traditionally we pass around platform and rule flags separately, with the idea
that all platform flags must be emitted before all rule flags (which may
override them).

This diff introduces a composable tuple of the two sets of flags, such that the
elements are concatenated pair-wise when the tuple itself is concatenated.

The implementation also contains two implementation classes, a conventional one
based on lists, and another one based on abstract iterables. This allows for
fewer memory allocations as these flag objects are concatenated.

Test Plan: existing tests pass.

Reviewed By: andrewjcg

fb-gh-sync-id: a55096f
shipit-source-id: a55096f