Extract provided symbols from prebuilt_jar rules automatically for autodeps.
Summary:This means that the hand-maintained configuration in `.buckconfig` can be
completely removed for the Buck project (this is done as part of this revision).
Support for the configuration still exists as an escape hatch.

Test Plan:`PrebuiltJarSymbolsFinderTest` verifies the new behavior in `PrebuiltJarSymbolsFinder`.

Re-ran `buck autodeps` with the newer, emptier `.buckconfig`.
The only change to `BUCK.autodeps` is where a new dep is added for
something that was already in the `exported_deps` of the rule.
This is not an error: just a missed opportunity for optimization.

Reviewed By: Coneko

fb-gh-sync-id: 19a3ab7
shipit-source-id: 19a3ab7