Don't create new CxxPlatform for each rule
Summary:Previously if a compilation database rule or an infer rule was
built without an explicit platform flavor, a new `CxxPlatform`
object would be created for each rule. Change that to pass a
`CxxPlatform` into `CxxLibraryDescription` and reuse it for all
rules, like it is done for example for `CxxBinaryDescription`.

Test Plan: `buck test`

Reviewed By: yiding

fb-gh-sync-id: af32cf0
shipit-source-id: af32cf0