Easy: Ignore BUCK_TTY when checking for changes to environment
Summary:In `Set env var when Buck is run from a TTY`, andrewjcg
added a new feature to allow Buck to log whether or not it was
connected to a TTY.

Since this feature used an environment variable, changes in that
variable were causing buckd to dump its parser cache.

This adds `$BUCK_TTY` to the list of ignored enviroment variables.
We don't filter it since it's used for logging.

Test Plan: CI

Reviewed By: ryu2

fb-gh-sync-id: 4832e5f
shipit-source-id: 4832e5f