Remove the AbstractDescriptionArg
There's no need to inherit from a base class for constructor
args. Doing so prevents Buck users from using domain classes
as an arg (which is, admittedly, a little odd but still nice
to have as a possibility)

Instead, add the parameter available on all args to the same
place as every other global parameter.