Moved server connect to another thread. Added connect logging
Summary: This diff moves the client connecting to the server on another thread, since before it blocked the UI thread. Also, one of the main problems of the buck plugin is that it can't connect to the buck server, so I've added logging for it.

Test Plan: Ran the plugin in intellij.

fb-gh-sync-id: 76ed7c0
shipit-source-id: 76ed7c0