DirCache to keep all temp files in temp folder so it is not involved in clean up process
Summary:Previously, dir cache held all tmp files in the root of the cache folder. During clean up, which could happen on any thread, it could attempt to clean
the temp file. Because of racing between cleaning thread and thread that owns that temp file, temp file is likely to moved into the place by the moment
when clean operation will start. Cleaning thread could attempt to delete temp file, but it won't be there (it will be moved already), and that will
procude NoSuchFile exception. Even though it won't break anything (it will just print it into console and proceed), this diff fixes this issue.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: d3b3a54
shipit-source-id: d3b3a54