Use shorter flavor names for cxx rules
Summary:Instead of using file path in the cxx rule flavor, only use its
last element. This will shorten the directory names used for
rules' outputs, which are currently gettting very close to the
filesystem limit. This should be safe because we still append a
hash of the full path to the flavor and it should not be too
confusing because we still use the file name in the flavor.

Test Plan: CI.

Reviewed By: Coneko

fb-gh-sync-id: bf46338
shipit-source-id: bf46338