Make the [tools] section .buckconfig documentation use the Soy template
Summary:When I added the documentation for the agentpath option in tools I
just copied the documentation style from the previous section on thift. However,
I should have used to existing Soy template.
This diff modifies the existing options in the [tools] section to use the Soy template was well as updated the
section I previously added for the agentpath to use the template.

Test Plan: Run a local document server and observe output.

Reviewed By: sdwilsh, bhamiltoncx

fb-gh-sync-id: c626bed
shipit-source-id: c626bed