Make ProjectBuildFileParserPool thread-safe
Summary:The getAllRulesAndMetaRules would only work if all calls to it
(and close) would come from the same thread (which is currently
the case) because there is a race between scheduling the work
on the executor, it actually executing and the future being added
to the pendingWork collection.
This also makes sure the parserRequest 'retry' future is re-scheduled
on the executor instead of running synchronously on the same thread.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: 2f2a222
shipit-source-id: 2f2a222