Add more verbose exceptions for malformed raw data in the parser.
Summary:We received a report of an NPE that could only be caused by a raw node
not containing a name or base path. Unfortunately the issue does not
reproduce and based on the stack alone it's not possible to say how
we ended up in a corrupt state. Hopefully dumping the contents of
the node will let us gain insight.
This also adds checks to verify the base path coming from the raw node
data aligns with the base path under it was cached.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: 1dc62d7
shipit-source-id: 1dc62d7