Add resource union options to AndroidResource.
Summary:Provide an option to union an android_resource with its deps. This results in a merged R.txt file and allows for integrating buck with libraries that rely heavily on resource merging to reference dependent library resources.
Closes https://github.com/facebook/buck/pull/653

Test Plan:Add unit tests to MiniAppt, RDotTxtEntry, and AaptResourceCollector

this commit resolves #643

Reviewed By: marcinkosiba

fb-gh-sync-id: 0170afe
shipit-source-id: 0170afe