Add a visibility option
In my current use case, after running the importer, I would have
had to dig through about twenty BUCK files and add a rule to
`visibility`. With this extra change, you can specify a set of rules
either in the JSON or on the command line, and they are added to the
visibilty of the artifacts specified.