Allow for Go global vendor-style directories
Summary: Add an option in .buckconfig for a global directory to be added to the import map for external dependencies similar to the /vendor directories found in the path, but used for all go projects.

Test Plan: Created a sample BUCK project with a vendor_path set in .buckconfig and a dependency inside that path.

Reviewed By: Coneko

fb-gh-sync-id: 4586964
shipit-source-id: 4586964