C/C++: include compiler tool deps in preprocess-only rules
Summary:We add the `CompilerDelegate` to the rule key of preprocess-only
C/C++ rules, so the input-based rule key builder expects to see
any associated deps in the dep list.  Previously, we only allowed
paths as the compiler tool, but now that we accept outputs generated
by other rules, this started failing.

Test Plan: update unittest

Reviewed By: yiding

fb-gh-sync-id: 20721da
shipit-source-id: 20721da