Kill subprocesses when nailgun connection dies
Summary:This diff makes Buck kill all its subprocesses, both when buckd dies
suddenly (e.g., with SIGKILL) and when a nailgun client exits.
It works by making a PTY, then making the PTY slave the Buck process
controlling tty.  When Buck launches a subprocess (either with
NuProcess or with the regular java.lang.Process), the subprocess
inherits the controlling terminal and process group of buckd.
If buckd dies, the FD that refers to the PTY master dies, which causes
the kernel to send SIGHUP to all members of the foreground process
group (i.e., buckd's children), which makes them die.

The other case is nailgun client exit.  Here, we don't actually use
the PTY at all.  Instead, we just send SIGHUP to our own process
group.  Since buckd and all its children share the process group, we
send SIGHUP to everyone.  Ordinarily, we (buckd) would be killing
ourselves along with all our children, but we set SIGHUP to `getpid`
during initialization, so we don't do die.  (Because `getpid` does nothing.)

But when we exec a child, the kernel resets any signal not set to `SIG_IGN` to `SIG_DFL`, so all our children, atomically, get their `SIGCHLD` handlers set to `SIG_DFL` at the moment of creation. Consequently, when either our process or the kernel sends `SIGHUP` to our process group, all our children die.

Now, you might say something like "`getpid` is not a signal handler!" or object that "getpid takes zero arguments" or say something about "undefined behavior", but consider this: on both OS X and Linux, on all supported ABIs, calling a function with extra arguments is harmless --- the called function ignores its extra arguments. `getpid` is also marked as async-signal-safe, so of course it's safe to run in the context of a signal handler. While `getpid` does return a value, it's safe in all the ABIs we support to "return" a value this way from a `void` function (like a signal handler ) --- the return-value register is caller-saved, and the value that `getpid` returns is ignored. Combining all these ideas, it's clear that `getpid` is acceptable to use *as* a signal handler. `getpid` is just a convenient async-signal-safe function that does nothing and is not `SIG_IGN`.

Why use `getpid` and not a JNA callback? Because JNA callbacks call into Java, and Java is not async-signal-safe. We need to use an existing C function that JNA can locate, and `getpid` appears to be the best choice.

Test Plan: attach strace to buckd; watch as all subprocesses die gloriously with SIGHUP after SIGINT kills the nailgun client

Reviewed By: bhamiltoncx

fb-gh-sync-id: 3ed6f20
shipit-source-id: 3ed6f20