New approach to capture only
Summary:The approach that used metadata to do capture-only doesn't work properly when it comes to creating
rules for public header symlinks.
This new approach is building an analysis-like graph as usual, but at the end it attaches nodes differently
to parallelize all the captures, since no analysis is needed.

Test Plan: Integration tests, CI

Reviewed By: Coneko

fb-gh-sync-id: 7108c2c
shipit-source-id: 7108c2c