Do not share the WorkerProcesses map when duplicating an ExecutionContext
Summary:When an ##ExecutionContext## closes, it calls ##close()## on all of its ##WorkerProcess## instances. When a copy of an ##ExecutionContext## was made, the copy would reference the same instances of ##WorkerProcess##. This was problematic because the ##WorkerProcess## instances would be closed twice, once when the copy context closed and once when the original context closed.

This diff fixes it, so that a copied ##ExecutionContext## is given an empty ##WorkerProcess## map. That way, we'll ensure a ##WorkerProcess## is only referenced by its original ##ExecutionContext##, and therefore only closed once.

Test Plan: buck build buck

Reviewed By: beefon

fb-gh-sync-id: 532eea6
shipit-source-id: 532eea6