Update ReactNative rules to use WorkerShellStep
Summary:This diff adds support for having ##ReactNativeDeps## and ##ReactNativeBundle## interact with a JS packager that implements the ##WorkerProcess## protocol. To enable this mode, you need to add ##use_worker true## and define ##packager_worker## in the ##react-native## section of the ##.buckconfig##:
  [react-native]
    packager_worker=/path/to/worker
    use_worker=true

Test Plan:Updated the integration tests to run the test cases with both the old packager interface and with the new worker interface.
  buck test //test/com/facebook/buck/js:js

Reviewed By: beefon

fb-gh-sync-id: dcd4629
shipit-source-id: dcd4629