Massive BUCK file cleanup.
Summary:Before `buck autodeps` would generate `exported_deps` for us,
we had to manually specify `exported_deps` when something in
the `srcs` would extend/implement a type from a dependency, or
for all sorts of other reasons.

Now that `buck autodeps` does this for us, we can remove all sorts
of manual configuration that is now automated.

Test Plan: CI

Reviewed By: yiding

fb-gh-sync-id: 6a6c460
shipit-source-id: 6a6c460