Use new BUCK.autodeps format that includes both deps and exported_deps.
Summary:This revision was generated by running `buck autodeps` on top of the
stack with D3074589 at the top of it.

Again, the only reason this is separate from D3074589 is to make
D3074589 easier to review. These two revisions must be landed together.

Test Plan: Unlike D3074589, CI should succeed at this revision.

Reviewed By: yiding

fb-gh-sync-id: ab06b35
shipit-source-id: ab06b35