C/C++: refactor `CxxHeaders` to wrap/encapsulate a single include root
Summary:Previously, we distributed information about the header pack presented
by the header symlink tree across the `CxxPreprocessorInput` and
`CxxHeaders` classes, and treated the latter as a glorified map of all
headers associated with the `CxxPreprocessorInput` object.

This diff refactors this a bit so that we now the `CxxHeaders` object
encapsulates all information about a header pack at a given root.
This means the information about the header symlink tree is completely
hidden/contained by `CxxHeaders`, and that each `CxxPreprocessorInput`
object may have multiple `CxxHeaders` objects (this isn't the case now,
but it will be when adding the new `header_dirs` parameter).

This also means that the `CxxHeaders.concat()` method has been dropped
in this diff, as it now doesn't make sense to combine multiple `CxxHeaders`
objects (since it must have a single root).  As such, the conflicting
header detection has been moved the `PreprocessorDelegate` class.

Test Plan: CI

Reviewed By: yiding

fb-gh-sync-id: 432a125
shipit-source-id: 432a125