add config option tools.java
Summary:Follow a pattern similar to tools.javac to add a config option tools.java

Eventually, the `Java` object can become a full-fledged tool, but for this diff
I went for a shorter path to sanely centralizing its path configuration.

Note that if tools.java is unset in the config, the behavior is same as before.

Test Plan:Self-reviewed the code for the following patterns:

* `*Step` classes hold onto a `Java` reference without `AddToRuleKey`
* `BuildRule` classes hold onto a `Java` reference *with* `AddToRuleKey`
* `*Description` classes hold onto a `JavaOptions` reference without `AddToRuleKey`

Made sure buck tests pass:

  $ ant clean test

Made sure it works end to end:

  $ cd ~/buck/ && ant && touch build/successful-build
  $ cd ~/test-project && ~/buck/bin/buck --config=tools.java=/usr/local/XXX test test-directory/...

Reviewed By: marcinkosiba

fb-gh-sync-id: 7c8e4a7
shipit-source-id: 7c8e4a7