Fix a bug in `buck project` to correctly handle genrule in java_library's srcs
Summary:Buck has a bug where if a java_library's srcs contains a genrule output, the IntellIJ module for that java_library will not contain the file generated by genrule. This has caused trouble when we moved some fragment model generation into Buck (see https://fb.facebook.com/groups/graphqlandroid/permalink/1020533541328594/ and https://fb.facebook.com/groups/askbuck/permalink/1021953111186637/). Because of that we had to roll back the feature.

This diff should fix the problem. Project.java now checks if one of the srcs's Path starts with buck-out and if it does, it adds that path to sourceFolders. You can also see how it works in the new unit test.

Test Plan: Added a unit test to cover this case.

Reviewed By: ASCE1885

fb-gh-sync-id: 5a89deb
shipit-source-id: 5a89deb