Log stderr to console if process closes unexpectedly
Summary: When a worker process throws an exception or unexpectedly closes, the error messaging is not helpful since it manifests as a JSON parsing error from the stacktrace. This diffs makes it so when we detect a failure in the external worker process, we'll log to the buck console any available data from the process's stderr stream before the build fails. This should make it a lot easier to figure out what is wrong with your worker process as you develop it.

Test Plan:Put a bad param in the ##worker_tool## for the JS packager, which causes it to fail.
Built a target that uses that ##worker_tool##.
Saw that buck printed the stderr to the console:
{F60448897}

Reviewed By: beefon

fb-gh-sync-id: b7fda2f
fbshipit-source-id: b7fda2f